+++ src/plotting/__init__.py
@@ -5,6 +5,9 @@ from plotting.plot_utils import (
     create_phase_plot,
     create_solution_plot,
     create_surface_plot,
+    create_vector_animation_3d,
+    create_vector_animation_plot,
+    export_animation_to_mp4,
     save_plot,
 )
 
@@ -13,5 +16,8 @@ __all__ = [
     "create_phase_plot",
     "create_solution_plot",
     "create_surface_plot",
+    "create_vector_animation_3d",
+    "create_vector_animation_plot",
+    "export_animation_to_mp4",
     "save_plot",
 ]